separate database client, entities from main
