+++ db/tables.go
@@ -1,0 +1,24 @@
+package main
+
+import (
+	"github.com/shopspring/decimal"
+	"gorm.io/gorm"
+)
+
+// AdItem is a table which contains information about all items
+type AdItem struct {
+	gorm.Model
+	ID           int `sql:"AUTO_INCREMENT" gorm:"primary_key"`
+	Title        string
+	Description  string
+	Price        decimal.Decimal `sql:"type:decimal(20,8);"`
+	ImageURLs    []ImageURL      `gorm:"foreignKey:AdItemID;constraint:OnUpdate:CASCADE,OnDelete:SET NULL;"`
+	MainImageURL *ImageURL       `gorm:"foreignKey:AdItemID;constraint:OnUpdate:CASCADE,OnDelete:SET NULL;"`
+}
+
+// ImageURL is a table for image URLs for items
+type ImageURL struct {
+	gorm.Model
+	URL      string
+	AdItemID int
+}
