+++ go.mod
@@ -1,4 +1,4 @@
-module golang--test-task
+module golang-test-task
 
 go 1.18
