+++ main.go
@@ -2,7 +2,8 @@ package main
 
 import (
 	"encoding/json"
-	"fmt"
+	"golang-test-task/database"
+	"golang-test-task/entities"
 	"io"
 	"net/http"
 	"net/url"
@@ -15,19 +16,10 @@ import (
 	"gorm.io/gorm"
 )
 
-const paginationSize int = 10
-
 type universalHandler struct {
-	DB        *gorm.DB
+	dbClient  *database.Client
 	validator *validator.Validate
 	logger    *zap.Logger
-}
-
-// Pagination contains all information required to paginate records' table
-type Pagination struct {
-	Offset int    `json:"offset"`
-	By     string `json:"by"`
-	Asc    bool   `json:"asc"`
 }
 
 // ListAds is a function to get list of ads
@@ -35,11 +27,11 @@ type Pagination struct {
 // sorting by price/date_created; asc/desc order
 // TODO: add pagination size to params
 func (u *universalHandler) ListAds(w http.ResponseWriter, r *http.Request) {
-	result := ListAdsAnswer{Status: "error"}
+	result := entities.ListAdsAnswer{Status: "error"}
 	if r.Method != "GET" {
 		w.WriteHeader(http.StatusBadRequest)
-		bs1, _ := json.Marshal(result)
-		_, _ = w.Write(bs1)
+		bs, _ := json.Marshal(result)
+		_, _ = w.Write(bs)
 		return
 	}
 
@@ -47,53 +39,34 @@ func (u *universalHandler) ListAds(w http.ResponseWriter, r *http.Request) {
 	if err != nil {
 		u.logger.Error("error during ReadAll in ListAds", zap.Error(err))
 		w.WriteHeader(http.StatusInternalServerError)
-		bs1, _ := json.Marshal(result)
-		_, _ = w.Write(bs1)
+		bs, _ := json.Marshal(result)
+		_, _ = w.Write(bs)
 		return
 	}
 
-	var pag Pagination
+	var pag entities.Pagination
 	err = json.Unmarshal(bs, &pag)
 	if err != nil {
 		u.logger.Error("error during Unmarshal in ListAds", zap.Error(err))
 		w.WriteHeader(http.StatusInternalServerError)
-		bs1, _ := json.Marshal(result)
-		_, _ = w.Write(bs1)
+		bs, _ = json.Marshal(result)
+		_, _ = w.Write(bs)
 		return
 	}
 
 	// TODO: make paginationSize customizable
-	items, err := u.listAds(pag.Offset, paginationSize, pag.By, pag.Asc)
+	items, err := u.dbClient.ListAds(pag.Offset, entities.PaginationSize, pag.By, pag.Asc)
 	if err != nil {
-		u.logger.Error("error during listAds in ListAds", zap.Error(err))
+		u.logger.Error("error during dbClient.ListAds in ListAds", zap.Error(err))
 		w.WriteHeader(http.StatusInternalServerError)
-		bs1, _ := json.Marshal(result)
-		_, _ = w.Write(bs1)
+		bs, _ = json.Marshal(result)
+		_, _ = w.Write(bs)
 		return
 	}
 
 	result.Result = items
-	bs1, _ := json.Marshal(result)
-	_, _ = w.Write(bs1)
-}
-
-func (u *universalHandler) listAds(offset, paginationSize int, by string, asc bool) (resItems []AdAPIListItem, err error) {
-	items := []*AdAPIListItem{}
-	ascOrDesc := "asc"
-	if !asc {
-		ascOrDesc = "desc"
-	}
-	order := fmt.Sprintf("%s %s", by, ascOrDesc)
-	db := u.DB.Preload("ImageURLs").Preload("MainImageURL").Model(&AdItem{}).Limit(paginationSize).Offset(offset).Order(order).Find(&items)
-	err = db.Error
-	if err != nil {
-		u.logger.Error("error during getting data in listAd", zap.Error(err))
-		return resItems, err
-	}
-	for _, v := range items {
-		resItems = append(resItems, *v)
-	}
-	return resItems, nil
+	bs, _ = json.Marshal(result)
+	_, _ = w.Write(bs)
 }
 
 // GetAd is a function to get concreate ad
@@ -101,11 +74,11 @@ func (u *universalHandler) listAds(offset, paginationSize int, by string, asc bo
 // required fields: title, price, main_photo_url
 // additional: by parameter `fields`(description, photo_urls)
 func (u *universalHandler) GetAd(w http.ResponseWriter, r *http.Request) {
-	result := GetAdAnswer{Status: "error"}
+	result := entities.GetAdAnswer{Status: "error"}
 	if r.Method != "GET" {
 		w.WriteHeader(http.StatusBadRequest)
-		bs1, _ := json.Marshal(result)
-		_, _ = w.Write(bs1)
+		bs, _ := json.Marshal(result)
+		_, _ = w.Write(bs)
 		return
 	}
 
@@ -113,88 +86,37 @@ func (u *universalHandler) GetAd(w http.ResponseWriter, r *http.Request) {
 	if err != nil {
 		u.logger.Error("error during ReadAll in GetAd", zap.Error(err))
 		w.WriteHeader(http.StatusInternalServerError)
-		bs1, _ := json.Marshal(result)
-		_, _ = w.Write(bs1)
+		bs, _ = json.Marshal(result)
+		_, _ = w.Write(bs)
 		return
 	}
 
-	var api GetAdAPI
+	var api entities.GetAdAPI
 	err = json.Unmarshal(bs, &api)
 	if err != nil {
 		u.logger.Error("error during Unmarshal in GetAd", zap.Error(err))
 		w.WriteHeader(http.StatusInternalServerError)
-		bs1, _ := json.Marshal(result)
-		_, _ = w.Write(bs1)
+		bs, _ = json.Marshal(result)
+		_, _ = w.Write(bs)
 		return
 	}
 
-	item, err := u.getAd(api.ID)
+	item, err := u.dbClient.GetAd(api.ID)
 	if err != nil && err != gorm.ErrRecordNotFound {
-		u.logger.Error("error during getAd in GetAd", zap.Error(err))
+		u.logger.Error("error during dbClient.GetAd in GetAd", zap.Error(err))
 		w.WriteHeader(http.StatusInternalServerError)
-		bs1, _ := json.Marshal(result)
-		_, _ = w.Write(bs1)
+		bs, _ = json.Marshal(result)
+		_, _ = w.Write(bs)
 		return
 	}
 	result.Status = "success"
 	if item != nil {
-		m := createMapFromAdItem(*item, api.Fields)
+		m := item.CreateMapFromAdItem(api.Fields)
 		result.Result = &m
 	}
 
-	// TODO: add item handling
-	// fmt.Println(item)
-	bs1, _ := json.Marshal(result)
-	_, _ = w.Write(bs1)
-}
-
-func createMapFromAdItem(item AdItem, fields []string) (m map[string]interface{}) {
-	m = map[string]interface{}{}
-	m["id"] = item.ID
-	m["title"] = item.Title
-	var u *string
-	if item.MainImageURL != nil {
-		// fmt.Println(item)
-		u = &item.MainImageURL.URL
-	}
-	// fmt.Println(item)
-	m["main_image_url"] = u
-	for _, field := range fields {
-		if field == "description" {
-			m["description"] = item.Description
-		} else if field == "image_urls" {
-			imgUrls := make([]string, 0)
-			for _, v := range item.ImageURLs {
-				imgUrls = append(imgUrls, v.URL)
-			}
-			m["image_urls"] = imgUrls
-		}
-	}
-	return m
-}
-
-// GetAdAPI stores information about interesting item and which fields to show
-type GetAdAPI struct {
-	ID     int      `json:"id"`
-	Fields []string `json:"fields"`
-}
-
-// GetAdAnswer couples a status of processing and a request's result
-type GetAdAnswer struct {
-	Status string                  `json:"status"`
-	Result *map[string]interface{} `json:"result"`
-}
-
-func (u *universalHandler) getAd(id int) (res *AdItem, err error) {
-	// TODO: add fields to use in .Select(fields)
-	var item AdItem
-	db := u.DB.Preload("ImageURLs").Preload("MainImageURL").First(&item, id).Association("MainImageURL")
-	err = db.Error
-	if err != nil {
-		u.logger.Error("error during getting data in getAd", zap.Error(err))
-		return nil, err
-	}
-	return &item, nil
+	bs, _ = json.Marshal(result)
+	_, _ = w.Write(bs)
 }
 
 // CreateAd is a function to create ad
@@ -202,11 +124,11 @@ func (u *universalHandler) getAd(id int) (res *AdItem, err error) {
 // Params: title, description, photo_urls, price
 // Return: ID of new ad, code of a result
 func (u *universalHandler) CreateAd(w http.ResponseWriter, r *http.Request) {
-	result := CreateAdAnswer{ID: nil, Status: "error"}
+	result := entities.CreateAdAnswer{ID: nil, Status: "error"}
 	if r.Method != "POST" {
 		w.WriteHeader(http.StatusBadRequest)
-		bs1, _ := json.Marshal(result)
-		_, _ = w.Write(bs1)
+		bs, _ := json.Marshal(result)
+		_, _ = w.Write(bs)
 		return
 	}
 
@@ -214,128 +136,42 @@ func (u *universalHandler) CreateAd(w http.ResponseWriter, r *http.Request) {
 	if err != nil {
 		u.logger.Error("error during ReadAll in CreateAd", zap.Error(err))
 		w.WriteHeader(http.StatusInternalServerError)
-		bs1, _ := json.Marshal(result)
-		_, _ = w.Write(bs1)
+		bs, _ = json.Marshal(result)
+		_, _ = w.Write(bs)
 		return
 	}
 
-	var item AdJSONItem
+	var item entities.AdJSONItem
 	err = json.Unmarshal(bs, &item)
 	if err != nil {
 		u.logger.Error("error during Unmarshal in CreateAd", zap.Error(err))
 		w.WriteHeader(http.StatusInternalServerError)
-		bs1, _ := json.Marshal(result)
-		_, _ = w.Write(bs1)
+		bs, _ = json.Marshal(result)
+		_, _ = w.Write(bs)
 		return
 	}
 	err = u.validator.Struct(item)
 	if err != nil {
 		u.logger.Error("error during validating in CreateAd", zap.Error(err))
 		w.WriteHeader(http.StatusBadRequest)
-		bs1, _ := json.Marshal(result)
-		_, _ = w.Write(bs1)
+		bs, _ = json.Marshal(result)
+		_, _ = w.Write(bs)
 		return
 	}
-	// for _, e := range err.(validator.ValidationErrors) {
-	//	fmt.Println(e)
-	//}
 
-	id, err := u.createAd(item)
+	id, err := u.dbClient.CreateAd(item)
 	if err != nil {
-		u.logger.Error("error during createAd in CreateAd", zap.Error(err))
+		u.logger.Error("error during dbClient.CreateAd in CreateAd", zap.Error(err))
 		w.WriteHeader(http.StatusInternalServerError)
-		bs1, _ := json.Marshal(result)
-		_, _ = w.Write(bs1)
+		bs, _ = json.Marshal(result)
+		_, _ = w.Write(bs)
 		return
 	}
 	result.Status = "success"
 	result.ID = &id
 
-	bs1, _ := json.Marshal(result)
-	_, _ = w.Write(bs1)
-}
-
-// CreateAdAnswer stores a status of created item
-// and an ID of created item(if it was created)
-type CreateAdAnswer struct {
-	Status string `json:"status"`
-	ID     *int   `json:"id"`
-}
-
-// ListAdsAnswer combine a status of a list process and a result
-type ListAdsAnswer struct {
-	Status string          `json:"status"`
-	Result []AdAPIListItem `json:"result"`
-}
-
-// ImageURL is a table for image URLs for items
-type ImageURL struct {
-	gorm.Model
-	URL      string
-	AdItemID int
-}
-
-// AdJSONItem contains the information to write to AdItem table
-type AdJSONItem struct {
-	Title       string          `json:"title" validate:"required,min=1,max=200"`
-	Description string          `json:"description" validate:"required,max=1000"`
-	Price       decimal.Decimal `json:"price" validate:"required,numeric"`
-	ImageURLs   []string        `json:"imageURLs" validate:"required,max=3,checkURL"`
-}
-
-// AdItem is a table which contains information about all items
-type AdItem struct {
-	gorm.Model
-	ID           int `sql:"AUTO_INCREMENT" gorm:"primary_key"`
-	Title        string
-	Description  string
-	Price        decimal.Decimal `sql:"type:decimal(20,8);"`
-	ImageURLs    []ImageURL      `gorm:"foreignKey:AdItemID;constraint:OnUpdate:CASCADE,OnDelete:SET NULL;"`
-	MainImageURL *ImageURL       `gorm:"foreignKey:AdItemID;constraint:OnUpdate:CASCADE,OnDelete:SET NULL;"`
-}
-
-// AdAPIListItem stores information is needed for pagination show
-type AdAPIListItem struct {
-	ID    int
-	Title string
-	Price decimal.Decimal
-}
-
-func (u *universalHandler) createAd(item AdJSONItem) (id int, err error) {
-	var mainImageURL *ImageURL
-	size := 0
-	imgURLsSize := len(item.ImageURLs)
-	if imgURLsSize > 0 {
-		mainImageURL = &ImageURL{URL: item.ImageURLs[0]}
-		if imgURLsSize > 1 {
-			size = imgURLsSize - 1
-		}
-	}
-
-	var imageURLs = make([]ImageURL, size)
-	if imgURLsSize > 1 {
-		arr := item.ImageURLs[1:]
-		for i := range arr {
-			if i == 0 {
-				continue
-			}
-			imageURLs[i].URL = arr[i]
-		}
-	}
-
-	fmt.Println("!", mainImageURL)
-
-	ad := AdItem{Title: item.Title, Description: item.Description, Price: item.Price,
-		ImageURLs: imageURLs, MainImageURL: mainImageURL,
-	}
-
-	db := u.DB.Create(&ad)
-	err = db.Error
-	if err != nil {
-		u.logger.Error("error during creating datum in createAd", zap.Error(err))
-		return 0, err
-	}
-	return ad.ID, nil
+	bs, _ = json.Marshal(result)
+	_, _ = w.Write(bs)
 }
 
 func main() {
@@ -357,26 +193,12 @@ func main() {
 		return true
 	})
 
-	// refer https://github.com/go-sql-driver/mysql#dsn-data-source-name for details
-	// dsn := "user:pass@tcp(127.0.0.1:3306)/dbname?charset=utf8mb4&parseTime=True&loc=Local"
-	// db, err := gorm.Open(mysql.Open(dsn), &gorm.Config{})
-
-	// c := zap.NewDevelopmentConfig()
-	// c.EncoderConfig.EncodeLevel = zapcore.CapitalColorLevelEncoder
-	// l, _ := c.Build()
-	// defer l.Sync()
-	// loggerG := zapgorm2.New(l)
-	// loggerG.SetAsDefault() // optional: configure gorm to use this zapgorm.Logger for callbacks
-
 	dsn := "host=postgres user=gorm password=gorm dbname=gorm port=5432 sslmode=disable TimeZone=Asia/Shanghai"
-	// db, err := gorm.Open(postgres.Open(dsn), &gorm.Config{Logger: loggerG})
-	db, err := gorm.Open(postgres.Open(dsn), &gorm.Config{
-		//Logger: glogger.Default.LogMode(glogger.Warn),
-	})
+	db, err := gorm.Open(postgres.Open(dsn), &gorm.Config{})
 	if err != nil {
 		panic("failed to connect database")
 	}
-	err = db.AutoMigrate(&AdItem{}, &ImageURL{})
+	err = db.AutoMigrate(&database.AdItem{}, &database.ImageURL{})
 	if err != nil {
 		panic("failed to automigrate")
 	}
@@ -384,20 +206,12 @@ func main() {
 	config := zap.NewDevelopmentConfig()
 	config.EncoderConfig.EncodeLevel = zapcore.CapitalColorLevelEncoder
 	logger, _ := config.Build()
-
-	// logger, _ := zap.NewDevelopment()
 	defer func() {
 		_ = logger.Sync()
 	}()
 
-	// d, err := db.DB()
-	////d.Close()
-	// d.Ping()
-	// if err != nil {
-	//	fmt.Println(err)
-	//}
-
-	logic := universalHandler{DB: db, validator: v, logger: logger}
+	client := database.NewClient(db)
+	logic := universalHandler{dbClient: client, validator: v, logger: logger}
 
 	mux := http.NewServeMux()
 	mux.HandleFunc("/create_ad", logic.CreateAd)
